requirements.txt
web3>=6.0.0
requests>=2.28.0
python-dotenv>=0.20.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
aiohttp>=3.8.0
python-dateutil>=2.8.2
